0x03-MessagingApp-Django/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
celery==5.4.0
redis==5.0.1
amqp==5.2.0
kombu==5.3.5
drf-yasg==1.21.7
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
